created new DTOs and migrations. Also added a comment in readme

+++ NET/Models/Obra.cs
@@ -21,7 +21,7 @@ public class Obra
     [Required]
     public int Duracion { get; set; }
     [Required]
-    public int Precio { get; set; }
+    public decimal Precio { get; set; }
     [Required]
     public List<ButacaObra> ListaButacaObra { get; set; }
 
@@ -29,7 +29,7 @@ public class Obra
 
     public Obra() { }
 
-    public Obra(string titulo, string descripcion, DateTime diaObra, string imagen, string genero, int duracion, int precio)
+    public Obra(string titulo, string descripcion, DateTime diaObra, string imagen, string genero, int duracion, decimal precio)
     {
         Titulo = titulo;
         Descripcion = descripcion;
